requirement.txt
flask
flask_session
openai
python-dotenv
azure-cognitiveservices-speech
pandas